feat(examples): scope policy to exact DeFi interactions; rename to with-aave-atomic-batch
createPolicy.ts now allowlists precisely the contracts (USDC, Aave Pool,
faucet) and function selectors (mint, approve, supply, borrow, repay,
withdraw) the example uses. Verified live that eth.tx.* predicates evaluate
against EVERY call in a V2 batch with all-or-nothing semantics:

- faucet / enter / exit batches: allowed and executed
- USDC.transfer (allowed contract, banned selector): denied
- call to an unlisted address: denied
- batch mixing one valid approve with one unlisted call: denied as a unit

README policy section updated accordingly (replaces the stale "per-call
predicates are evolving" caveat).
