+++ examples/defi/with-aave-atomic-batch/src/exit.ts
