+++ examples/defi/with-aave-atomic-batch/src/faucet.ts
