+++ examples/defi/with-aave-atomic-batch/tsconfig.json
